+++ packages/tailwindcss-language-server/tests/document-links/document-links.test.js
@@ -131,6 +131,44 @@ withFixture('v4/basic', (c) => {
     ],
   })
 
+  testDocumentLinks('source not: file exists', {
+    text: '@source not "index.html";',
+    lang: 'css',
+    expected: [
+      {
+        target: `file://${path
+          .resolve('./tests/fixtures/v4/basic/index.html')
+          .replace(/@/g, '%40')}`,
+        range: { start: { line: 0, character: 12 }, end: { line: 0, character: 24 } },
+      },
+    ],
+  })
+
+  testDocumentLinks('source not: file does not exist', {
+    text: '@source not "does-not-exist.html";',
+    lang: 'css',
+    expected: [
+      {
+        target: `file://${path
+          .resolve('./tests/fixtures/v4/basic/does-not-exist.html')
+          .replace(/@/g, '%40')}`,
+        range: { start: { line: 0, character: 12 }, end: { line: 0, character: 33 } },
+      },
+    ],
+  })
+
+  testDocumentLinks('@source inline(…)', {
+    text: '@source inline("m-{1,2,3}");',
+    lang: 'css',
+    expected: [],
+  })
+
+  testDocumentLinks('@source not inline(…)', {
+    text: '@source not inline("m-{1,2,3}");',
+    lang: 'css',
+    expected: [],
+  })
+
   testDocumentLinks('Directories in source(…) show links', {
     text: `
       @import "tailwindcss" source("../../");